refactor: 검정, 흰색 점수 분리
