+++ README.md
@@ -13,6 +13,12 @@
 - GameSession
     - [x] 현재 시스템 내에서 실행되고 있는 게임을 관리한다.
 
+- BoardSession
+
+- [x] 현재 시스템에서 돌아가고 있는 보드를 관리한다.
+- [x] 보드가 존재하는지 확인한다.
+- [x] 보드를 세션에 등록한다.
+
 #### 컨트롤러
 
 - GameCommand
@@ -148,9 +154,9 @@
         - [x] 한 진영에 왕이 없으면 해당 진영은 0점이다.
         - [x] 왕 혼자 있더라도 해당 진영은 0점이다.
 
-#### Score
+#### WhiteScore
 
-- [x] 보드 데이터와 색을 통해 점수를 계산한다.
+- [x] 보드 내 흰색 기물의 점수이다.
     - [x] 왕이 죽었으면 점수는 0점이다.
     - [x] 왕은 점수 갖지않는다.
     - [x] Queen 은 9점으로 계산한다.
@@ -160,11 +166,17 @@
     - [x] Pawn 은 1점으로 계산한다.
         - [x] 같은 세로줄에 있는 폰은 0.5점으로 계산한다.
 
-#### BoardSession
+#### BlackScore
 
-- [x] 현재 시스템에서 돌아가고 있는 보드를 관리한다.
-- [x] 보드가 존재하는지 확인한다.
-- [x] 보드를 세션에 등록한다.
+- [x] 보드 내 검은색 기물의 점수이다.
+    - [x] 왕이 죽었으면 점수는 0점이다.
+    - [x] 왕은 점수 갖지않는다.
+    - [x] Queen 은 9점으로 계산한다.
+    - [x] Rook 은 5점으로 계산한다.
+    - [x] bishop 은 3점으로 계산한다.
+    - [x] Knight 는 2.5점으로 계산한다.
+    - [x] Pawn 은 1점으로 계산한다.
+        - [x] 같은 세로줄에 있는 폰은 0.5점으로 계산한다.
 
 ### 뷰
