+++ src/main/java/chess/domain/game/Game.java
@@ -2,7 +2,6 @@ package chess.domain.game;
 
 import chess.domain.Position;
 import chess.domain.board.Board;
-import chess.domain.piece.Color;
 
 public abstract class Game {
     protected final Board board;
@@ -13,7 +12,7 @@ public abstract class Game {
     }
 
     public static Game of(Board board) {
-        if (board.getScoreOf(Color.WHITE).getScore() == 0 || board.getScoreOf(Color.BLACK).getScore() == 0) {
+        if (board.whiteScore().getScore() == 0 || board.blackScore().getScore() == 0) {
             return new EndGame(board);
         }
         return new RunningGame(board);
@@ -24,7 +23,7 @@ public abstract class Game {
     public abstract boolean isEnd();
 
     public Result getResult() {
-        return new Result(board.getScoreOf(Color.WHITE), board.getScoreOf(Color.BLACK));
+        return new Result(board.whiteScore(), board.blackScore());
     }
 
     public Board getBoard() {
